requirements.txt
# Django framework
Django==5.0.0

# WSGI server (if using WSGI)
Gunicorn>=20.0

# ASGI server (if using ASGI)
Uvicorn>=0.15
djangorestframework==3.15.1
psycopg2-binary==2.9.9
